moving mapping rules to the conversion folder
