+++ metricflow/conversion/dbt_mapping_rules/dbt_metric_model_to_data_source_rules.py
@@ -3,7 +3,7 @@ from typing import List, Tuple
 
 from dbt_metadata_client.dbt_metadata_api_schema import MetricNode
 from metricflow.dataflow.sql_table import SqlTable
-from metricflow.model.dbt_mapping_rules.dbt_mapping_rule import (
+from metricflow.conversion.dbt_mapping_rules.dbt_mapping_rule import (
     DbtMappingRule,
     MappedObjects,
     assert_metric_model_name,
