+++ metricflow/conversion/dbt_mapping_rules/dbt_metric_to_measure.py
@@ -3,7 +3,7 @@ from typing import Dict, List, Tuple
 
 from dbt_metadata_client.dbt_metadata_api_schema import MetricNode
 from dbt_semantic_interfaces.objects.aggregation_type import AggregationType
-from metricflow.model.dbt_mapping_rules.dbt_mapping_rule import (
+from metricflow.conversion.dbt_mapping_rules.dbt_mapping_rule import (
     DbtMappingRule,
     MappedObjects,
     assert_essential_metric_properties,
