+++ metricflow/conversion/dbt_mapping_rules/dbt_metric_to_metrics_rules.py
@@ -2,7 +2,7 @@ import traceback
 from typing import List, Tuple
 
 from dbt_metadata_client.dbt_metadata_api_schema import MetricFilter, MetricNode
-from metricflow.model.dbt_mapping_rules.dbt_mapping_rule import (
+from metricflow.conversion.dbt_mapping_rules.dbt_mapping_rule import (
     DbtMappingRule,
     MappedObjects,
     assert_essential_metric_properties,
