+++ metricflow/test/model/dbt_cloud_parsing/test_dbt_metric_to_measure.py
@@ -1,18 +1,18 @@
 from typing import Tuple
 
 from dbt_metadata_client.dbt_metadata_api_schema import MetricNode
-from metricflow.model.dbt_mapping_rules.dbt_metric_to_measure import (
+from metricflow.conversion.dbt_mapping_rules.dbt_metric_to_measure import (
     DbtToMeasureName,
     DbtToMeasureAgg,
     DbtToMeasureAggTimeDimension,
     DbtToMeasureExpr,
 )
-from metricflow.model.dbt_mapping_rules.dbt_mapping_rule import (
+from metricflow.conversion.dbt_mapping_rules.dbt_mapping_rule import (
     MappedObjects,
     DbtMappingRule,
     get_and_assert_calc_method_mapping,
 )
-from metricflow.model.dbt_converter import DbtConverter
+from metricflow.conversion.dbt_converter import DbtConverter
 from dbt_semantic_interfaces.objects.metric import MetricType
